requirements.txt
pyyaml
argparse
scipy
tqdm
tensorboardX
tensorboard
setproctitle
numpy
pytest-shutil